refactor(test_base.py): use `make_raw_request_message` for backward compatibility
